+++ CHANGELOG.md
@@ -7,7 +7,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## Unreleased
 ### Added
+- Added `getIndex` and `setIndex` to `QValue`. Strings can now be indexed.
 - `compile` function: Parses a string into a function object that can be called from within imquery.
+- `version` function: Get the version of imquery.
+- `char` function: Get a string from a character code.
+- `find` function: find the position of a string
 
 ### Changed
 - Lambda block syntax now does away with the `=>`, while blockless syntax keeps it.